Moved PreferencesViewModel
