+++ app/src/main/kotlin/fobo66/exchangecourcesbelarus/di/ViewModelsModule.kt
@@ -19,7 +19,8 @@ package fobo66.exchangecourcesbelarus.di
 import dev.fobo66.valiutchik.presentation.MainViewModel
 import dev.fobo66.valiutchik.presentation.OpenSourceLicensesViewModel
 import dev.fobo66.valiutchik.presentation.OpenSourceLicensesViewModelImpl
-import fobo66.exchangecourcesbelarus.ui.preferences.PreferencesViewModel
+import dev.fobo66.valiutchik.presentation.PreferencesViewModel
+import dev.fobo66.valiutchik.presentation.PreferencesViewModelImpl
 import org.koin.core.module.dsl.viewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
@@ -29,5 +30,7 @@ val viewModelsModule = module {
     viewModel<OpenSourceLicensesViewModel> {
         OpenSourceLicensesViewModelImpl(get())
     }
-    viewModelOf(::PreferencesViewModel)
+    viewModel<PreferencesViewModel> {
+        PreferencesViewModelImpl(get(), get(), get(), get())
+    }
 }
