+++ app/src/main/kotlin/fobo66/exchangecourcesbelarus/ui/Navigation.kt
@@ -42,12 +42,12 @@ import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
 import dev.fobo66.valiutchik.presentation.MainViewModel
 import dev.fobo66.valiutchik.presentation.OpenSourceLicensesViewModel
+import dev.fobo66.valiutchik.presentation.PreferencesViewModel
 import dev.fobo66.valiutchik.presentation.entity.MainScreenState
 import fobo66.exchangecourcesbelarus.R
 import fobo66.exchangecourcesbelarus.ui.licenses.OpenSourceLicensesScreen
 import fobo66.exchangecourcesbelarus.ui.main.BestRatesGrid
 import fobo66.exchangecourcesbelarus.ui.preferences.PreferenceScreen
-import fobo66.exchangecourcesbelarus.ui.preferences.PreferencesViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import org.koin.androidx.compose.koinViewModel
